+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0](https://github.com/pgerke/homebridge-freeathome-local-api/releases/tag/1.1) - 03.07.2022
+
+### Added
+
+- [PG-192](https://pgerke.atlassian.net/browse/PG-192):
+  Support door openers
+- [PG-203](https://pgerke.atlassian.net/browse/PG-203):
+  Allow the user to specify channels and devices to ignore
+
 ## [1.0.0](https://github.com/pgerke/homebridge-freeathome-local-api/releases/tag/1.0) - 27.06.2022
 
 This is the first official release. The plugin now has full code coverage, meets the defined quality gate with respect to code style and has passed extensive stability tests as well as static code analysis, so I feel confident ending the preview phase at this time. The release version retains all features from v0.4.0-pre.